feat: 랜딩페이지 리다이렉트 재구현
